+++ plugins/plotly-express/src/deephaven/plot/express/data/__init__.py
@@ -8,4 +8,8 @@ from .data_generators import (
     wind,
     gapminder,
     fish_market,
+    outages,
+    flights,
+    OUTAGE_CENTER,
+    FLIGHT_CENTER,
 )